requirements.txt
## 🔧 Configuración de Versiones Compatibles

### Versiones Críticas para Compatibilidad
```txt
# Core ML and Computer Vision
tensorflow==2.16.1
keras==3.11.3
mediapipe==0.10.21
opencv-python==4.8.1.78
numpy==1.26.4
pandas==2.0.3

# Data Processing
scikit-learn==1.6.1
joblib==1.3.2

# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pillow==10.0.1
```

### ⚠️ Notas Importantes de Compatibilidad
- **TensorFlow 2.16.1** es requerido para cargar los modelos entrenados
- **NumPy 1.26.4** es compatible con TensorFlow 2.16.1
- **scikit-learn 1.6.1** es necesario para los archivos .pkl
- **MediaPipe 0.10.21** es la versión estable más reciente

## �� Uso del Sistema

### Ejecución Básica
```bash
# Asegúrate de tener el entorno virtual activado
venv\Scripts\activate

# Ejecutar el sistema
python main.py
```

### Comandos de Ayuda
```bash
# Mostrar ayuda
python main.py --help
python main.py -h
python main.py help
```

### Controles Durante la Ejecución
- **Q** - Salir del programa
- **R** - Reiniciar sistema
- **S** - Capturar pantalla
- **H** - Mostrar ayuda

## 📊 Información del Modelo

### Modelos Disponibles
1. **Dense_Simple_patient.h5**
   - Precisión: 98.75%
   - Arquitectura: Red neuronal densa simple
   - Características: 258 (33 pose × 4 + 21 mano derecha × 3 + 21 mano izquierda × 3)
   - Clases: 30

2. **final_correct_model.h5**
   - Precisión: 98.75%
   - Arquitectura: Modelo optimizado
   - Características: 258
   - Clases: 30

### Clases de Lenguaje de Señas Reconocidas